Birth of the InventoryManager
I originally intended to start my development on FuelManager first, but
then realized I needed to use the Inventory to do that, so I figured,
why not start with the Inventory?
Anyway, InventoryManager is currently a minimalistic stub, with none of
the Turtle Function Overloads intended to make it magical.